[tfman] move remaining parsing code into parse.c
